Seperate tui.py into seperate files
