+++ microproxy/viewer/tui/__init__.py
@@ -1,0 +1,1 @@
+from tui import start
