+++ microproxy/viewer/tui/style.py
@@ -1,0 +1,9 @@
+palette = [
+    ("header", "white", "dark green", "bold"),
+    ("footer", "white", "dark blue", "bold"),
+    ("message", "white", "black"),
+    ("message focus", "black", "light gray"),
+    ("prop", "white", "black"),
+    ("prop focus", "black", "light gray"),
+    ("prop separator", "white", "dark red", "bold")
+]
