+++ microproxy/viewer/tui/widgets/__init__.py
@@ -1,0 +1,1 @@
+from parent_frame import ParentFrame
